Consolidate Market Reporter routines
